+++ languagetool-core/src/main/java/org/languagetool/JLanguageTool.java
@@ -765,6 +765,7 @@ public class JLanguageTool {
     thisMatch.setSuggestedReplacements(match.getSuggestedReplacements());
     thisMatch.setUrl(match.getUrl());
     thisMatch.setSynonymsFor(match.getSynonymsFor());
+    thisMatch.setType(match.getType());
     String sentencePartToError = sentence.substring(0, match.getFromPos());
     String sentencePartToEndOfError = sentence.substring(0, match.getToPos());
     int lastLineBreakPos = sentencePartToError.lastIndexOf('\n');
@@ -1113,6 +1114,7 @@ public class JLanguageTool {
             newMatch.setEndColumn(range.to.column);
             newMatch.setSuggestedReplacements(match.getSuggestedReplacements());
             newMatch.setSynonymsFor(match.getSynonymsFor());
+            newMatch.setType(match.getType());
             adaptedMatches.add(newMatch);
           }
           ruleMatches.addAll(adaptedMatches);