+++ languagetool-core/src/main/java/org/languagetool/rules/AbstractDateCheckFilter.java
@@ -87,7 +87,9 @@ public abstract class AbstractDateCheckFilter extends RuleFilter {
               .replace("{realDay}", getDayOfWeek(dateFromDate))
               .replace("{day}", getDayOfWeek(calFromDateString))
               .replace("{currentYear}", Integer.toString(Calendar.getInstance().get(Calendar.YEAR)));
-      return new RuleMatch(match.getRule(),match.getSentence(), match.getFromPos(), match.getToPos(), message, match.getShortMessage());
+      RuleMatch ruleMatch = new RuleMatch(match.getRule(), match.getSentence(), match.getFromPos(), match.getToPos(), message, match.getShortMessage());
+      ruleMatch.setType(match.getType());
+      return ruleMatch;
     } else {
       return null;
     }