+++ languagetool-core/src/main/java/org/languagetool/rules/ShortenedYearRangeChecker.java
@@ -43,7 +43,7 @@ public class ShortenedYearRangeChecker extends RuleFilter {
       String centuryPrefix = arguments.get("x").substring(0, 2);
       int y = Integer.parseInt(centuryPrefix + arguments.get("y"));
       if (x >= y) {
-        return new RuleMatch(match.getRule(), match.getSentence(), match.getFromPos(), match.getToPos(), match.getMessage(), match.getShortMessage());
+        return match;
       }
     } catch (IllegalArgumentException ignore) {
       // if something's fishy with the number – ignore it silently,