+++ languagetool-core/src/main/java/org/languagetool/rules/YMDDateHelper.java
@@ -18,7 +18,6 @@
  */
 package org.languagetool.rules;
 
-import java.util.Calendar;
 import java.util.Map;
 
 /**
@@ -52,7 +51,9 @@ public class YMDDateHelper {
     String correctDate = String.format("%d-%s-%s", correctYear, month, day);
     String message = match.getMessage()
             .replace("{realDate}", correctDate);
-    return new RuleMatch(match.getRule(), match.getSentence(), match.getFromPos(),
+    RuleMatch ruleMatch = new RuleMatch(match.getRule(), match.getSentence(), match.getFromPos(),
             match.getToPos(), message, match.getShortMessage());
+    ruleMatch.setType(match.getType());
+    return ruleMatch;
   }
 }