+++ languagetool-core/src/main/java/org/languagetool/rules/spelling/hunspell/HunspellRule.java
@@ -116,6 +116,7 @@ public class HunspellRule extends SpellingCheckRule {
             len, len + word.length(),
             messages.getString("spelling"),
             messages.getString("desc_spelling_short"));
+        ruleMatch.setType(RuleMatch.Type.UnknownWord);
         if (userConfig == null || userConfig.getMaxSpellingSuggestions() == 0 || ruleMatches.size() <= userConfig.getMaxSpellingSuggestions()) {
           List<String> suggestions = getSuggestions(word);
           List<String> additionalTopSuggestions = getAdditionalTopSuggestions(suggestions, word);